+++ apps/web/src/shared/hooks/useVisitorId.ts
@@ -1,39 +1,21 @@
-import { useEffect, useState, useCallback, useRef } from "react";
-import { api } from "../lib/api-client";
+"use client";
 
-/** localStorageキー: visitorId */
-export const VISITOR_ID_KEY = "recommend_scp_visitor_id";
-/** localStorageキー: オンボーディング完了フラグ */
-export const ONBOARDING_COMPLETED_KEY = "recommend_scp_onboarding_completed";
+import { useContext } from "react";
+import { VisitorContext } from "../contexts/VisitorProvider";
+import type { VisitorContextValue } from "../contexts/VisitorProvider";
+
+// 定数をre-export（既存の利用箇所との後方互換性を維持）
+export { VISITOR_ID_KEY, ONBOARDING_COMPLETED_KEY } from "../contexts/VisitorProvider";
 
 /** useVisitorIdの戻り値の型 */
-export interface UseVisitorIdResult {
-  /** 現在のvisitorId（未初期化時はnull） */
-  visitorId: string | null;
-  /** 初期化中またはAPI呼び出し中 */
-  isLoading: boolean;
-  /** オンボーディング完了済みかどうか */
-  isOnboarded: boolean;
-  /** エラー情報 */
-  error: Error | null;
-  /** visitorIdをリフレッシュ（デバッグ用） */
-  refresh: () => Promise<void>;
-}
-
-/** API レスポンスの型 */
-interface VisitorResponse {
-  visitorId: string;
-  isNew: boolean;
-  createdAt: string;
-  onboardingCompletedAt?: string | null;
-}
+export type UseVisitorIdResult = VisitorContextValue;
 
 /**
- * visitorIdを管理するカスタムフック
+ * visitorIdを取得するカスタムフック
  *
- * - localStorageにvisitorIdが存在しない場合、新規UUIDを生成
- * - 常にAPIを呼び出してDBと同期（visitorが存在しない場合は自動作成）
- * - オンボーディング完了状態も管理
+ * VisitorProvider内で使用すること。
+ * Providerが提供するContextから値を読み取るため、
+ * 複数箇所で呼び出してもAPIリクエストは1回のみ。
  *
  * @example
  * ```tsx
@@ -45,78 +27,11 @@ interface VisitorResponse {
  * ```
  */
 export function useVisitorId(): UseVisitorIdResult {
-  const [visitorId, setVisitorId] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-  const [isOnboarded, setIsOnboarded] = useState(false);
-  const [error, setError] = useState<Error | null>(null);
+  const context = useContext(VisitorContext);
 
-  // アンマウント検知用
-  const isMountedRef = useRef(true);
+  if (!context) {
+    throw new Error("useVisitorId must be used within a VisitorProvider");
+  }
 
-  const initialize = useCallback(async () => {
-    setIsLoading(true);
-    setError(null);
-
-    try {
-      // 1. localStorageから取得、または新規UUID生成
-      const storedVisitorId = localStorage.getItem(VISITOR_ID_KEY);
-      const visitorIdToUse = storedVisitorId ?? crypto.randomUUID();
-
-      // 2. APIに登録/同期（既存visitorIdでもAPIを呼んでDBと同期）
-      // これによりDBにvisitorが存在しない場合も自動的に作成される
-      const res = await api.visitors.$post({
-        json: { visitorId: visitorIdToUse },
-      });
-
-      // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- res.ok は実行時に false になる可能性がある
-      if (!res.ok) {
-        throw new Error("Failed to register visitor");
-      }
-
-      const data: VisitorResponse = await res.json();
-
-      // 3. 状態を更新（アンマウント済みの場合はスキップ）
-      if (isMountedRef.current) {
-        // localStorageに保存
-        localStorage.setItem(VISITOR_ID_KEY, visitorIdToUse);
-
-        // サーバーからonboardingCompletedAtが返された場合は同期
-        if (data.onboardingCompletedAt) {
-          localStorage.setItem(ONBOARDING_COMPLETED_KEY, "true");
-          setIsOnboarded(true);
-        } else {
-          // サーバーにonboardingCompletedAtがない場合はlocalStorageも同期
-          const storedOnboardingCompleted = localStorage.getItem(ONBOARDING_COMPLETED_KEY);
-          setIsOnboarded(storedOnboardingCompleted === "true");
-        }
-
-        setVisitorId(visitorIdToUse);
-      }
-    } catch (e) {
-      if (isMountedRef.current) {
-        setError(e instanceof Error ? e : new Error("Unknown error"));
-      }
-    } finally {
-      if (isMountedRef.current) {
-        setIsLoading(false);
-      }
-    }
-  }, []);
-
-  useEffect(() => {
-    isMountedRef.current = true;
-    void initialize();
-
-    return () => {
-      isMountedRef.current = false;
-    };
-  }, [initialize]);
-
-  return {
-    visitorId,
-    isLoading,
-    isOnboarded,
-    error,
-    refresh: initialize,
-  };
+  return context;
 }
